Add ability to work within container. Added more logging and error-checking.

+++ devcontainer/Extensions.cs
@@ -1,9 +1,14 @@
+using System.Xml.Linq;
+using Anotar.LibLog;
+using System.Text;
+using System.Diagnostics;
 using System;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Collections;
 using System.Collections.Generic;
+
 
 namespace devcontainer
 {
@@ -54,7 +59,7 @@ namespace devcontainer
                 File.WriteAllText(destFilename, text);
         }
 
-        public static void CopyTo(this string sourceFolder, string destinationFolder, string mask = "*.*", bool createFolders = false, bool recurseFolders = false, bool overwrite = false, 
+        public static void CopyTo(this string sourceFolder, string destinationFolder, string mask = "*.*", bool createFolders = true, bool recurseFolders = true, bool overwrite = false, 
             Action<string, string> onCopyFile = null, bool overWrite = false)
         {
             try
@@ -65,24 +70,72 @@ namespace devcontainer
 
                 foreach (string sourceFile in Directory.GetFiles(dir.ToString(), mask, so))
                 {
-                    FileInfo srcFile = new FileInfo(sourceFile);
+                    var srcFile = new FileInfo(sourceFile);
                     string srcFileName = srcFile.Name;
 
                     var destFile = new FileInfo(destinationFolder + srcFile.FullName.Replace(sourceFolder, ""));
 
                     if (!Directory.Exists(destFile.DirectoryName ) && createFolders)
+                    {
                         Directory.CreateDirectory(destFile.DirectoryName);
+                        LogTo.Info($"Creating folder {destFile.DirectoryName}...");
+                    }
 
                     if (onCopyFile == null)
-                        File.Copy(srcFile.FullName, destFile.FullName, overwrite);
+                    {
+                        LogTo.Info("Using default copy");
+                        if (File.Exists(destFile.FullName) && !overWrite)
+                        {
+                            LogTo.Warn($"Destination file {destFile.FullName} exists and over-writing is forbidden. Skipping.");
+                            continue;
+                        }
+                        else
+                        {
+                            LogTo.Info($"Copying {srcFile.FullName} -> {destFile.FullName}");
+                            File.Copy(srcFile.FullName, destFile.FullName, overwrite);
+                        }
+                    }
                     else
+                    {
+                        LogTo.Info($"Using custom copy for {srcFile.FullName} -> {destFile.FullName} ...");
                         onCopyFile(srcFile.FullName, destFile.FullName);
+                    }
                 }
             }
             catch (Exception ex)
             {
                 Console.Error.WriteLine($"Exception: {ex.Message}\n{ex.StackTrace}");
             }
+        }
+
+        public static string EnsureDirectoriesExist(this string path)
+        {   
+            Directory.CreateDirectory(path);
+            return path;
+        }
+
+        public static void WriteEnvFile(this IDictionary<string, string> env, string filename)
+        {
+            var text = new StringBuilder();
+            foreach (var kvp in env)
+                text.AppendLine($"{kvp.Key}={kvp.Value}");
+            File.WriteAllText(filename, text.ToString());
+        }
+        public static void AppendToEnvFile(this IDictionary<string, string> env, string filename)
+        {
+            var text = new StringBuilder();
+            foreach (var kvp in env)
+                text.AppendLine($"{kvp.Key}={kvp.Value}");
+            File.AppendAllText(filename, text.ToString());
+        }
+
+        public static bool RunningInContainer()
+        {
+            var filename = $"/proc/1/cgroup";
+            if (!File.Exists(filename))
+                return false;
+            var text = File.ReadAllText(filename);
+            return text.Contains("/docker/");
         }
     }
 }
